+++ src/redux/services/compare/transferResource.ts
@@ -1,3 +1,5 @@
+import {KubernetesObject} from '@kubernetes/client-node';
+
 import {cloneDeep} from 'lodash';
 import {v4 as uuid} from 'uuid';
 
@@ -7,7 +9,7 @@ import {AppDispatch} from '@models/appdispatch';
 import {K8sResource} from '@models/k8sresource';
 import {RootState} from '@models/rootstate';
 
-import {ResourceSet} from '@redux/reducers/compare';
+import {ResourceSet} from '@redux/compare';
 import {currentConfigSelector, kubeConfigContextSelector, kubeConfigPathSelector} from '@redux/selectors';
 import {applyResource} from '@redux/thunks/applyResource';
 import {updateResource} from '@redux/thunks/updateResource';
@@ -86,8 +88,15 @@ async function deployResourceToCluster(
 
   // Remark: Cluster adds defaults so copying the source's content
   // is too naive. Instead fetch remotely and fallback to copy if failed.
-  const clusterContent = await getResourceFromCluster(source, kubeConfigPath, currentContext);
-  const updatedContent = clusterContent ?? source.content;
+  let updatedContent: KubernetesObject;
+  try {
+    const sourceCopy = structuredClone(source);
+    sourceCopy.namespace = namespace;
+    const clusterContent = await getResourceFromCluster(sourceCopy, kubeConfigPath, currentContext);
+    updatedContent = clusterContent ?? source.content;
+  } catch {
+    updatedContent = source.content;
+  }
 
   const id = target?.id ?? uuid();
   const resource = createResource(updatedContent, {