+++ src/redux/store.ts
@@ -4,10 +4,11 @@ import {createLogger} from 'redux-logger';
 
 import {sectionBlueprintMiddleware} from '@src/navsections/sectionBlueprintMiddleware';
 
+import * as compareListeners from './compare/listeners';
+import {compareSlice} from './compare/slice';
 import {combineListeners, listenerMiddleware} from './listeners/base';
 import {alertSlice} from './reducers/alert';
 import {configSlice} from './reducers/appConfig';
-import {compareListener, compareSlice, filterListener, resourceFetchListener} from './reducers/compare';
 import {extensionSlice} from './reducers/extension';
 import {logsSlice} from './reducers/logs';
 import {imageSelectedListener, mainSlice, resourceMapChangedListener} from './reducers/main';
@@ -29,10 +30,10 @@ if (process.env.NODE_ENV === `development`) {
 export const resetStore = createAction('app/reset');
 
 combineListeners([
-  resourceFetchListener('left'),
-  resourceFetchListener('right'),
-  compareListener,
-  filterListener,
+  compareListeners.resourceFetchListener('left'),
+  compareListeners.resourceFetchListener('right'),
+  compareListeners.compareListener,
+  compareListeners.filterListener,
   resourceMapChangedListener,
   imageSelectedListener,
 ]);