+++ src/redux/thunks/previewKustomization.ts
@@ -2,6 +2,7 @@ import {createAsyncThunk} from '@reduxjs/toolkit';
 
 import log from 'loglevel';
 import path from 'path';
+import {v4 as uuid} from 'uuid';
 
 import {ERROR_MSG_FALLBACK, ROOT_FILE_ENTRY} from '@constants/constants';
 
@@ -14,7 +15,7 @@ import {currentConfigSelector} from '@redux/selectors';
 import {getK8sVersion} from '@redux/services/projectConfig';
 import {createPreviewResult, createRejectionWithAlert} from '@redux/thunks/utils';
 
-import {CommandResult, hasCommandFailed, runCommandInMainThread} from '@utils/command';
+import {CommandResult, hasCommandFailed, runCommandInMainThread} from '@utils/commands';
 import {DO_KUSTOMIZE_PREVIEW, trackEvent} from '@utils/telemetry';
 
 /**
@@ -104,6 +105,7 @@ export function runKustomize(
   }
 
   return runCommandInMainThread({
+    commandId: uuid(),
     cmd: projectConfig?.settings?.kustomizeCommand ? String(projectConfig.settings.kustomizeCommand) : 'kubectl',
     args,
     env: {