+++ src/utils/kubeclient.ts
@@ -1,10 +1,11 @@
 import * as k8s from '@kubernetes/client-node';
 
 import log from 'loglevel';
+import {v4 as uuid} from 'uuid';
 
 import {AppConfig, ClusterAccess, ClusterAccessWithContext, KubePermissions} from '@models/appconfig';
 
-import {runCommandInMainThread} from '@utils/command';
+import {runCommandInMainThread} from '@utils/commands';
 import electronStore from '@utils/electronStore';
 import {getMainProcessEnv} from '@utils/env';
 
@@ -110,6 +111,7 @@ function parseCanI(stdout: string, namespace: string): ClusterAccess {
 export async function getKubeAccess(namespaces: string[], currentContext: string): Promise<ClusterAccessWithContext[]> {
   const promises = namespaces.map(namespace => {
     return runCommandInMainThread({
+      commandId: uuid(),
       cmd: 'kubectl',
       args: ['auth', 'can-i', '--list', `--namespace=${namespace}`],
     });